Location Like
Added code to like media posted in a location. Works on multiple locations as well, though getting the location-id is a bit tricky you need to manually search for it and then you can use that, ill try to understand how the location system works in instagram and if i can workout there system i would post it but until then i think manual way is the only solution.

P.s im not a coder just used some common sense to work it all out so if the code breaks im really sorry. Tested on python 3 working flawlessly for me